SmrGalaxy: don't store escaped name in database
This continues the effort to store raw data in the database, leaving
it intact to be processed in whatever ways are necessary (since an
escaped string cannot be "unescaped").

We add `SmrGalaxy::getDisplayName` and replace most calls to `getName`
with this new method. Also remove `SmrSector::getGalaxyName` as it is
a wrapper that becomes less useful with this design.

I confirmed that no galaxy names in the live database have the `&`
character, so no database changes should be needed.