Change how TIME is modified in CLI programs
We use a constant TIME throughout the code to ensure that a single
page request has a consistent time associated with it (time changing
over the course of a page load could cause logical errors).

The CLI programs are long-running scripts that make many "quasi" page
requests (e.g. NPC actions), and so TIME has to be modified. Since
PHP constants cannot be modified natively, we were using the UOPZ
extension to modify them. This is a big tool for a small job, and so
it makes sense to replace it with something simpler.

We add a Time class and store the time (in seconds and microseconds)
there. The SmrSession is then responsible for instantiating it and
accessing its values. The Time object is recreated if we want to
update to a new time (e.g. in the CLI programs), and we throw an
exception if this is attempted in normal page requests.

NOTE: Since `TIME` was previously a constant expression, it could be
used in class attribute initialization. This is no longer the case,
which affected at least one class (AbstractSmrPort).