Use a single Dockerfile for all CLI programs (#943)
The discord, irc, and npc CLI programs use very similar environments,
but were built into separate Docker images. We now combine them all
into a single "CLI" image.

The combined image is a bit larger than it needs to be for any one of
these programs, but reducing the number of built images significantly
reduces local/DockerHub build resource costs (which is especially
important for reducing the turnaround time when pushing to the `live`
branch for production). It also reduces repetition of Dockerfile and
composer.json boilerplate.