add scouts to backyard drone farm (#949)
Currently scouts are only sold at a Combat Accessories during normal games.  This makes scouts unnecessarily rare as the CA locations are tightly controlled in order to keep mines away from fed beacons.  Furthermore, placing a CA in a place accessible to racial galaxies but not in a place to be mined in is a time consuming pain in the ass.

We could bypass that SNAFU by simply adding scout drones to the drone farm inventory.  Which this commit does!  (assuming no errors were committed!)