Always create new instances of MySqlDatabase (#951)
Fixes a bug with overlapping queries clobbering previous results.

The old SmrMySqlDatabase had a non-static `$dbResult`, so any call to
`new SmrMySqlDatabase()` would have a new `$dbResult`. However, the
original design of `MySqlDatabase::getInstance` from #924 shared a
`$dbResult` (due to the shared MySqlDatabase from the DI container),
which caused query results to be overwritten if another query was made
while processing the results of the first query. The global variable
`$db` is kind of what `MySqlDatabase::getInstance` was supposed to be.

We now only use `getNewInstance` (which has been renamed `getInstance`
to reduce the footprint of the diff). To handle the reconnection logic
for CLI programs, we set the `mysqli` instance in the DI container to
null, and then reconnect in the MySqlDatabase ctor if it is passed a
null `$dbConn` (since there is no known way to "unset" an entry in the
DI container so that it is regenerated on-demand the next time it is
requested).

Co-authored-by: Dan Hemberger <daniel.hemberger@gmail.com>