Fix edge case when resizing galaxies
We were only modifying sectors if the galaxy size changed. However,
this neglected the edge case where the dimensions change, but the
total number of sectors is unchanged.

In this case, we still need to update the `sector` table because the
sector neighbors will change. The galaxy dimension modification code
does more than is needed here, but it is safe to use it.