SmrPlayer: fix bug in leaveAlliance
This bug was introduced in 7efea2361f0, by calling `isAllianceLeader`
after the player's alliance ID was set to 0. As a result, the player
would never be identified as the alliance leader, and thus incur a
delay even when leaving a 1-player alliance (which was not intended).

No practical impact because the alliance join delay was set to 0 in
c33cc7d5f54da4. However, it must be fixed now because it throws an
error when using typed class properties.