Improve fault tolerance when adding locations
A database error was being triggered when adding a location that
already existed in the sector. This is unnecessarily strict, and we
can instead simply skip adding locations that already exist.

This fix is easily achieved by holding the locations to add in an
associative array with location ID keys. We can also simplify the
logic -- no need to differentiate between "keep" and "add" if we
modify `addLocationToSector` to appropriately skip adding linked
locations if they already exist.