Remove the CLI Docker image
The php-apache image used for the main game engine can also be used
for CLI programs. Even though this image is much heavier than we need
for the CLI programs (it was nice having a minimal alpine image), it
didn't provide any tangible benefits due to Docker's copy-on-write
strategy. Having fewer images to maintain is a very real benefit.