Fix the test script timing out on Travis
> [Symfony\Component\Process\Exception\ProcessTimedOutException]
> The process "docker-compose --env-file test.env run --rm smr-integration-
> test" exceeded the timeout of 300 seconds.

By default, Composer has a 300-second timeout for scripts. On Travis,
this was sometimes exceeded when running the `test` Composer script.
This was due in part to Travis not having a mechanism to cache Docker
builds, but it's also conceivable that the tests could eventually take
longer than 300 seconds to run.

Since Travis has a timeout mechanism of its own, we can simply remove
the Composer timeout entirely.