bootstrap.php: don't call exit during error handling
Using `exit` during `logException` makes this function unusable in
any context where processing needs to continue after the exception
is logged (note that this only impacted debug mode). We also remove
the return value because it is not needed.

In `handleException`, we also don't want `exit` for the same reasons,
but it is inconsequential in this case because there is not currently
any processing that takes place after this function.