Dockerfile: include patch-level version in FROM
Specifying the patch-level version in the FROM statement ensures that
we get a rebuild for a newer version even if the builder is using a
Docker cache (e.g. building locally or on DockerHub).

This fixes an issue with `pecl install` that was present in PHP 7.4.0,
but was fixed in a later version.