Prevent unnecessary page refresh errors
On pages with form submission, we typically save the form data in the
SmrSession $var. If the player then refreshes this page and resubmits
the form data (which is the default for most browsers), then we would
raise an exception due to the data being present both in $var and in
$_REQUEST. We want to allow this behavior, while still catching any
coding errors that result from setting session data incorrectly.

To do this, we now allow duplicate indices in $var and $_REQUEST if
and only if the values are the same in both. This could still hide a
coding error (if the values were accidentally the same), but at least
in this case it has no impact on the displayed page (unlike if the
values are different).