Add constants for trade transaction types
In the context of port transactions, we try to consistently use the
terms "Buy" and "Sell" from the perspective of the player. However,
it is not always clear which perspective is being used. To address
this deficiency, we make the following substitutions:

'Buy' => TRADER_BUYS
'Sell' => TRADER_SELLS

This should make the intended behavior unambiguous, but should also
make it much easier to reverse the perspective should we decide that
it's more intuitive the other way (i.e. TRADER_BUYS => PORT_SELLS,
etc.).