Change credits given to players on death
Previously, players would get 100k credits or 1/4 their ship value,
whichever was higher.

Now, players will get the starting credits + 1/4 their ship value.

This change allows us to modify the minimum credits supplied on death
for a given game (whereas previously it was a hard-coded constant).
However, it would be even better to have a separate knob to change
the on-death credits.

The two components are added instead of taking the larger amount only
for simplicity. It also makes death slightly less punishing (in a
monetary sense) for players in ships less than ~1M credits, which is
probably a good thing.