Display more accurate script runtime
There is a good chunk of time (~10-20 milliseconds) spent processing
prior to when SmrSession initializes the request time. To get a more
accurate total runtime, we can use `$_SERVER['REQUEST_TIME_FLOAT']`
instead. Note that this still excludes the time between when the
template is rendered and when it is displayed to the user.