Fix Globals::getAllianceNewsHREF regression
In 67658e8a, the expected var name in news_read_advanced.php changed
from `AllianceID` to `allianceID`. Since it is passed through the
SmrSession as an int, we also need to fetch it with `getRequestVarInt`
instead of `getRequestVar` (which is for string types only).

This fixes the "View News" link in the alliance menu.