Refactor .env file setup
Volume-mounting individual files is unreliable due to limitations
with the way Docker handles inode changes. If editing a file creates
a new inode, then those changes will unexpectedly not be propagated
into the container.

See https://github.com/moby/moby/issues/15793#issuecomment-135411504.

To fix this, we will avoid volume-mounting files. We move the `.env`
file to `config/env`, so that we can reuse the `config` directory
volume-mount. We also create a symlink to `.env` for the convenience
of docker-compose commands (since it reads `.env` by default).