docker-compose.yml: add connectRetries to flyway
From the Flyway docs:

> The maximum number of retries when attempting to connect to the
> database. After each failed attempt, Flyway will wait 1 second
> before attempting to connect again, up to the maximum number of
> times specified by connectRetries.

With `-connectRetries=20`, this allows us to run flyway even if the
mysql service takes up to 20 seconds to spin up. It is effectively
a replacement for the docker-compose v2 healthcheck, and a better
alternative to the more intrusive "wait-for-it" style bash scripts
(or, less elegantly, a `sleep 20`).