SmrPlayer: follow SQL table lock rules
When locking a table for writes, it is an error to _read_ from any
unlocked table in the same session (they must be locked as well).

Since `SmrSession::getTime()` can call `SmrSession::init` (a database
heavy function) if it is the first time SmrSession is loaded, we need
to ensure that we do not call it inside of a table lock.