run-tests.yml: use buildkit for docker builds
GitHub Actions environments support buildkit, which is a new build API
that is supposed to offer a lot of benefits over the legacy build
(see https://docs.docker.com/develop/develop-images/build_enhancements/).

We use the "plain" progress output so that the CI logs have the full
output (since the default buildkit output is compactified).