Add Google social login
We use The PHP League's Google OAuth2 provider.

Google Sign-In button is taken directly from the official assets at
https://developers.google.com/identity/branding-guidelines. The only
change is to scale the button to match the other social buttons
(6px of padding was removed, the "G" was scaled by 1.5x, and shifted
to the center of the icon).