shop_goods.php: fix illegal HTML in trade table
It is not legal HTML to put a `<form>` element inbetween various
table elements (it either needs to wrap the entire table or exist
entirely inside a `<td>` element).

Since each row of the trade table is its own form, and we have
inputs from multiple columns, this causes a problem for us. We see
the following HTML validator errors:

> Element “form” not allowed as child of element “tbody” in this
> context.

The best solution is to use the HTML5 `form` attribute, which can
be used to associate, e.g., an `<input>` to a `<form>` of which it
is not a child element. The `input.form` value must match the
`form.id` value.

Note that this HTML5 feature has 97.6% global browser compatibility,
and only notably excludes all IE browsers.