Template: no longer strip whitespace
Why is it okay to stop stripping whitespace? When content is served
with gzip compression, there is a negligible difference in bandwidth
expense.

Why would we want to stop stripping whitespace? Developing/debugging
is much harder when the HTML is on a single line. We could turn off
whitespace stripping when in debug mode, but then we'd need a second
stage of development where we verify that stripping whitespace results
in the same HTML output. Lastly, the code to strip whitespace is
complicated (it's not just a `trim`), so there is a maintenance cost
to keeping it around.