Improve combat result serialization
For SmrPlanet and SmrForce objects, we were serializing every class
property. However, this created larger and less portable serialized
data. By reducing the data to a small set of metadata (to, mostly,
only what is needed to display the combat result), we improve the
efficiency and portability of the combat result serialization.