SmrPlanet: change how exists() works
Add a new property `exists` that will be set during construction and
returned by the `exists()` method, rather than indirectly relying on
how the constructor sets the `gameID` and `sectorID` properties.

This will be especially helpful when type declarations are added, so
that we avoid complications with uninitialized typed properties.