Fix mysqli corruption in unit tests
In MySqlDatabaseTest.php we were setting a mock mysqli instance in the
DiContainer (which wasn't even being used), and this was causing any
integration tests run afterwards to fail all mysqli queries.

Since the MySqlDatabaseTest.php unit tests weren't using the mysqli
mock, it makes sense to merge them into the integration test and fore-
go the mock setup entirely.

It seems plausible that this file was accidentally commited, and was
an earlier development version of MySqlDatabaseIntegrationTest.php.