Require @covers annotations in phpunit
The @covers annotation specifies which parts of the code are tested
explicitly by a test. We require this annotation to help avoid false
positives for code coverage (i.e. testing a line of code implicitly).

Any test that is not designed to test a specific class should use the
@coversNothing annotation to indicate this so that it is not marked
as a risky test.