SmrPlanet: harden destroyBuilding
So long as you had at least 1 building of the given type, you could
destroy more of those buildings than you had, resulting in setting
a negative number of buildings.

To fix this, we move the sanity check into the lowest-level function,
`setBuildings`, and demand that we don't set a negative number.

Add some unit testing of the SmrPlanet building methods.