Reverse malformed assertions in unit tests
In PHPUnit, the syntax is `assert(expected, actual)`, and I had often
written assertions with the arguments reversed. Though the reversed
ordering seems more natural to me, we need to conform to the interface
as defined, since it is necessary to correctly interpret the results
of tests in some cases, e.g.

Failed asserting that two arrays are identical.
--- Expected
+++ Actual
@@ @@
-Array &0 (
-    4 => 2
-)
+Array &0 ()