login.css: fix whitespace in small menu
We stopped stripping whitespace in 83655c40a, which caused gaps to
between login page images that were supposed to be adjacent.

These gaps are fixed by moving the text formatting CSS into the child
divs that only contain the text (and not the imgs), because then the
imgs inherit `font-size: 0` from the parent using `collapsed`.