SmrAlliance: rework canJoinAlliance method
Replace `canJoinAlliance` with `getJoinRestriction`, which essentially
flips the boolean return value from `true` to `false` in the case where
there is not a join restriction.

This is helpful because `false` is a special type that is allowed only
in unions. Previously, we would have needed to specify `bool` because
`true` is not a valid return type (alone or in unions).