Remove redundant information in class docstrings
Documentation that non-programmatically repeats programmatic details
(such as the name of the class or the namespace that it's in) provide
no additional information, and can easily get out of sync. Therefore
it is best to simply omit it.

While @package is meaningful to phpDocumentor, it's more useful at
this point to categorize things by namespace only. If we need packages
in the future, we can add them.