BaseIntegrationSpec.php: remove verbose output
The printed message about connecting to MySQL interferes with the
default compact format of the PHPUnit output. It would be useful to
print this message if using `--debug` (which outputs individual test
names as they are run), but I cannot find any way to access this
option value within a PHPUnit\Framework\TestCase instance.