Refactor SmrShip for clearer database separation
The AbstractSmrShip class now contains _all_ logic for interacting
with a ship instance (and it is no longer `abstract`). The SmrShip
class now contains _only_ the logic for loading data from and writing
data to the database, based on how the properties in AbstractSmrShip
have changed.

This will allow us to test an AbstractSmrShip in its entirety without
any database access. It also simplifies the implementation of the
DummyShip class (since there are no more `abstract` methods it needs
to implement).