SmrShip: remove "old hardware" paradigm
We will no longer try to infer whether a ship is under attack by
keeping track of the "old hardware" values.

This was unnecessarily complicated, and the methods for setting the
"old hardware" were prone to misuse. They also caused inconsistency
in the display for the player (if two players are triggering on each
other, sometimes you will see the "under attack" warning, but other
times you won't).

Other problems include NPCs getting stuck in a state of thinking they
were under attack (which is kind of ironic...), and needing to call
the `removeUnderAttack` function in the many places where hardware is
being modified in a non-combat context.

The fix is to remove the "old hardware" paradigm entirely, and then
replace it with a simple `under_attack` column in the `player` db.
This is set explicitly _only_ when a player is actually under attack.

Gameplay Note: The "under attack" warning will now display whenever
you are the target of a weapon (whether it is fired defensively as
the result of your own trigger, or offensively as the result of
another player's trigger). There is potentially some tactical intel
lost because of this change, but it is reliable and unambiguous.