Dockerfile: conditionally install dev dependencies
Only install the composer dev dependencies when using `PHP_DEBUG=1`.
This will help reduce the size of the production image.

Also change the value of PHP_DEBUG in `env.sample` from 0 to 1, so
that users will install the dev dependencies by default (otherwise
they would not be able to run tests).