MySqlDatabase: fix repeated calls to close()
In #924, the `$dbConn` property was changed so that it is unset when
`close()` is called instead of being set to `false`. Therefore, when
we call `close()` on an already closed MySqlDatabase instance, we need
to perform an `isset` check on `$dbConn` instead of a boolean check.

This fix allows us to make repeated calls to `close()` again, and we
add a test to ensure this functionality.

Note that this impacted the IRC bot, which will make repeated calls to
close without reconnecting to the database if it processes an event
that doesn't require the database.