irc.php: fix PHP 8 error
As of PHP 8, it is a fatal error to pass a bool to `preg_replace`,
whereas in PHP 7.4, it returned an empty string.

This change broke irc.php, because we treated the result of `fgets`
as a string (by passing it to `preg_replace`) before we verified that
it was not `false`, which is the value it returns when there is no
message from the server (or in the case of an error).

To fix this, we simply rearrange the logic at the beginning of the
`readFromStream`. The checks independent of the server message are
performed first (1. server timeout check and 2. player timer check).
Next we poll for a message from the server, and at this point we can
properly return early if we get `false` (no message or error).

Setting the "last ping" time has been moved into the `server_ping`
function, which seems like a more appropriate place for it.
(Pings come about once every 90 seconds, and our timeout is 300s.)