run-tests.yml: split up build and test stage
Running these commands in separate stages will make it easier to find
the output of interest, since GitHub only provides structured output
at the stage level (and not individual commands within each stage).

The build output is quite long, so it takes a while to find the test
output when they are run in the same stage.