SmrShip: replace base ship with SmrShipType class
The `baseShip` property of SmrShip was just an array, which made it
difficult to validate. As a separate class, SmrShipType does not have
such a limitation.

In variable names, try to make it clearer when data is a ship versus
a ship type (e.g. `ship_id` -> `ship_type_id`).

The class structure is very similar to SmrWeaponType.

We provide an interface to get many ships in a single database query,
which is used in, e.g. SmrLocation::getShipsSold.

The following SmrShip methods have been modified:

* SmrShip::buildBaseShip -> SmrShipType::get
* SmrShip::getAllBaseShips -> SmrShipType::getAll
* SmrShip::{regenerateBaseShip -> regenerateShipType}
* SmrShip::{setShipTypeID -> setTypeID}
* SmrShip::{getShipTypeID -> getTypeID}
* SmrShip::{getShipClassID -> getClassID}
* SmrShip::getRefundValue (added)
* SmrShip::getMaxPower (moved to SmrShipType)
* SmrShip::getMaxHardware (moved to SmrShipType)
* SmrShip::canHave* (moved to SmrShipType)