SmrSession: prevent integer SNs
Previously, SNs were generated with a hash function that returned an
8-length hex string. Sometimes these strings were numeric, and since
the SN is used as an array key in `self::$vars`, it was implicitly
converted to an integer. From the PHP array manual:

> Additionally the following key casts will occur:
> * Strings containing valid decimal ints, unless the number is
>   preceded by a + sign, will be cast to the int type.

This is obviously incompatible with an interface that is expecting to
return a string.

To fix this, we generate the SN from the character set [a-z] with the
function `random_alphabetic_string` instead of [0-9a-f] from `md5`.

Note that this function has a cost linear with length (unlike `md5`,
which has a constant cost), and at a length of 8, it is about twice as
expensive (though still very fast: ~0.05ms per call). However, we do
not really need such a long string (2e11 combinations), so we reduce
it to a length of 6 (3e8 combinations). At most, we only store a few
hundred SNs in the session var for any given session.