Replace array containers with the Page class
The container mechanism for creating links and forwarding to new pages
originally used arrays. Now we use the Page class, which inherits from
ArrayObject. Since Objects (including ArrayObject) behave differently
for copying and references than arrays, this first stage of refactoring
only attempts to recover the old behavior without modifying the _logic_
of container usage everywhere in the code. Further stages may attempt
to treat containers as Objects from the ground up.

In particular, since we often reuse containers to make multiple HREFs,
we need to do a lot of (otherwise unnecessary) copying of the data in
the ArrayObject to prevent overwriting the data from prior usage.

The new interface maps fairly directly to the previous one:

* create_container(...) -> Page::create(...)
* $container = $var -> $container = Page::copy($var)
* SmrSession::getNewHREF($container) -> $container->href()
* forward($container) -> $container->go()
* resetContainer($container) -> $container->useAsGlobalVar()
* transfer(...) -> $container->addVar(...)

There are a number of subtle considerations in this refactoring. Here
is a mostly comprehensive list of them:

* Some of the container manipulation performed in SmrSession methods
  is now performed in the Page class (such as generating the CommonID
  and setting the ReaminingPageLoads.

* Functions that had container arguments can now be type-hinted as
  `Page` instead of the much more ambiguous `array`.

* We no longer use `$container` as a global variable. It's not clear
  why or if this was needed in the past, but it did not appear to be
  needed before or after this refactoring.

* Using `Page::copy($var)` as a static function instead of using a
  method (e.g. `$var->copy()`) will enable better static analysis,
  because the analyis will be able to infer the return type (whereas
  it doesn't know the type of the global variable `$var`).

* An undocumented difference between arrays and ArrayObjects is that
  `unset($var['non-existent key'])` emits an "Undefined array key" for
  ArrayObject, but not array. As a result, there were some places where
  `isset` checks needed to be performed where they weren't before.