Use MySqlDatabase::{get,escape}Object (#1032)
These methods are a wrapper around serialize and gzcompress for
database escaping and querying. Use them, where applicable, instead
of getting the raw strings and manually (un)serializing/compressing
them.

For the various attack pages, we can store the combat results in the
global `$var` in its raw form. Serializing them only to unserialize
them later just creates extra work.

This also includes two simplifications in SmrSession:

1. We always specify `active_session.session_var` on insertion, and
   `self::$var` is always set to an array; therefore, we know that
   the `session_var` can be safely extracted with `getObject`, rather
   than getting the raw string and verifying that it is not empty.

2. Since we don't always specify `active_session.ajax_returns`, it
   now has a default value of null. This allows us to use `getObject`
   without first checking if the raw string is the empty string.

(Note that the empty raw string is what you get when you have a column
default of `None`, but omit the field on insertion, which is not legal
in strict sql mode. We do not use this mode, but may in the future.)