Move engine processing from do_voodoo to Page class (#1035)
Add function `Page::process` that performs the processing of engine
pages. This helps to ensure a sanitized namespace that only has the
global variables in it that we expect to find. The goal is to remove
all of these global variables eventually (if for no other reason than
to improve static analysis of engine files).

Also remove special handling of error.php. We can use `require` for
every page instead of `require_once`.