Store time in DI Container instead of SmrSession
While it makes sense for the "page request time" to be stored in the
SmrSession in the traditional webserver context, it is too complex for
other purposes where the SmrSession doesn't make as much sense (e.g.,
unit testing, APIs, command-line programs).

Therefore, we add a set of static methods to the `Time` class (renamed
to `Epoch` due to perhaps unfounded fears about conflicting with the
`time` function since PHP is case-insensitive). These methods interact
with the DI container, which stores an instance of `Epoch` that will
be reused for every call to these static methods.

  SmrSession::getTime() -> Epoch::time()
  SmrSession::getMicroTime -> Epoch::microtime()
  SmrSession::updateTime -> Epoch::update()