SmrSession: remove USING_AJAX definition
This class had a spurious side effect when loaded in that it would
set `USING_AJAX=false` if it was not set yet. However, this variable
is set in config.php, which should _always_ be called as a prereq to
loading this class. Furthermore, loading classes should not have
spurious effects like this in general. Therefore, we remove it.