Page: remove Expires property
We never set the `Expires` property of a `Page` to a value greater
than 0, which means it was always ignored by `Session`. We base the
expiration of pages on the number of remaining page loads, not on an
absolute time.

Since we are not using this feature, we just remove it.