Globals: remove global player in getSector{Scan,Move}HREF
These functions require a `$player` so that they can determine the
target sector's relative direction from the player, which is then used
to set up the scan/move hotkeys.

There is probably a better way to organize the code so that generating
the HREF for sector scans/moves isn't responsible for setting hotkeys,
but there was no immediately obvious way to refactor this code.

In the meantime, we simply pass the current `$player` down through the
interface so that it is accessible to these lower level functions.

Also rename SmrSector::getScanSectorHREF to getSectorScanHREF for
consistency with the names of all the related functions.