trader_search_result.php: remove global player
Since the `playerLinks` function is specifically for displaying links
for the current player, we replace the global `$player` by getting it
from the Session (as opposed to passing it as an argument).

Also rename `$curr_player` to `$linkPlayer` to avoid confusion. It is
the player who is being searched for, _not_ the session player.