bootstrap.php: remove global account/player
In `logException`, get the player and account from the Session instead
of using globals. This is a bit less fault-tolerant, but I think we
need to assume some basic level of functioning for this logging to
work at all in the first place.

Also add "Game ID" to the output, since somehow that was missing from
the logging!