Menu: get player from the Session intstead of global
It's hard to decide when an object (e.g. player or account) should
be passed in as a parameter vs. internally taken from the Session.

In this case we'd never want to use anything other than the Session's
player to generate the menus, so it seemed wrong to make the caller
decide _which_ player to use.

Also: Fix council pages so that if you switch to view a different
race, clicking on "View Council" won't change back to the player's
race unexpectedly.