Remove $player and $account from global variables
Instead, these objects will be accessed from the Session. This will
facilitate improved static analysis of all engine files, since the
`$player` and `$account` object types will now be well-defined.

Note that because of the way these objects are stored in the Session,
dependency injection isn't really possible. In the future we may want
to instead store these objects in the DI container.

Additional notes:

* In `doSkeletonAssigns`, remove the conditional on the existence of
  an account in the Session, since we only ever use the skeleton.php
  after login.

* In error.php, we don't need to both check if the Session has a game
  ID _and_ if `$player` is an object, since they are redundant.