Add support for phpstan
PHPStan (or "PHP Static Analyzer") provides much more useful feedback
about the correctness of the code than the built-in PHP linter (i.e.
`php -l`, which we use in the linting unit test).

We reuse the docker-compose configuration from `smr-integration-test`
(which we refactor into the `x-smr-test` YAML anchor) and define a
basic set of options for PHPStan that should provide a good starting
point for triage.

Note that, beyond level 0, PHPStan reports a _lot_ of errors, so it
cannot immediately be used in CI. Some parts of the code will need to
be completely rewritten before this can be used in an automated way.

Only config.php is autoloaded presently, because the autoloaded files
are the only ones that are actually executed by PHP, and bootstrap.php
currently raises some confusing errors. Nevertheless, PHPStan still
can flag a large number of relevant code issues.

Explanation of other PHPStan options:

* `--memory-limit=4G` is used because otherwise the default memory
  limit of the PHP interpreter is used, which is much too low.

* `--ansi` is used to force color output, since running it through
  Composer makes it think the output isn't a TTY.