Rename MySqlDatabase to Smr\Database
Move this class into the `Smr` namespace so that it can be loaded by
the PSR-4 autoloader.

While some would argue that the "MySql" prefix is more accurate, and
they are _currently_ correct, we would prefer if the interface used by
the rest of the codebase was agnostic to the implementation details of
the Database class (in case we want to change it in the future), in
which case the more generic name is appropriate.

We also rename MySqlProperties to DatabaseProperties for consistency.