Let the DI Container manage the Template instance
Since we only have a single instance of the `Template` class in each
execution, it makes sense for it to be managed by the DI Container.
This allows us to access this singleton with `Template::getInstance`.

Ultimately, this will facilitate the removal of the `$template` global
variable.

We also move Template to the `Smr` namespace so that it can be loaded
by the PSR-4 autoloader.