Remove global $var from most functions
Use `SmrSession::getCurrentVar` instead.

While this seems like a simple change, it has a pretty big impact on
the inner workings of the Page and Session classes. This is because
they were treating the global `$var` and the underlying data in the
Session (`$this->var[$this->SN]`) as separate objects, which is a
vestigate of when vars were arrays instead of Page objects. Now that
they are objects, we can make modifications by reference to avoid the
"double updating".

This two major concept changes are as follows:

1. Updating data in the current var. Session::updateVar now only needs
   to update one object instead of two. However, that we technically
   do not need this method since the var can be modified by reference.

        $session->updateVar(name, value) -> $var[name] = value

2. Updating the entire current var to a new Page (i.e. forwarding).
   Page::useAsGlobalVar was responsible for updating the global `$var`
   and then Session::resetLink would update the internal Session data.
   The role of Page::useAsGlobalVar is no longer needed, so we remove
   it and call Session::resetLink (renamed `setCurrentVar`) directly.

   Note that when updating an element of the `$this->var` array, in
   order to retain references to an element (a Page object), we need
   to use `exchangeArray` to modify the contents in-place.