Page: allow changing index name in addVar
Previously, `Page::addVar` took a single string input, and insisted
that this be the index name in both `$var` and in the new container.

Now it takes two string inputs, `$source` and `$dest`. If `$dest` is
specified, this will be used as the index name in the new container.

This is useful when you want to transfer a value to a new container,
but want to give it a different name.