Session: define SN property in constructor
We were getting the SN in the Session constructor anyway, so it didn't
make sense to get it externally and pass it in via `findCurrentVar`.

There also doesn't seem to be any purpose to the `updateSN` function.
The `active_session.last_sn` field only matters on the next page load,
when we check `hasChangedSN`, and it is sufficient to update this when
the `update` method is called.