Session: refactor findCurrentVar method
Replace with `hasCurrentVar`, which does not modify the current var in
any way (i.e. it is a "const" method, unlike `findCurrentVar`). Thus,
it can safely be called any number of times, which is a useful query
to make before calling `getCurrentVar`.

The logic for modifying the var of the current SN that was previously
in `findCurrentVar` (de-decrementing RemainingPageLoads and unsetting
its CommonID) is moved to `fetchVarInfo`. However, instead of undoing
what `fetchVarInfo` did, we simply skip these operations for the var
associated with the current SN (which has the added benefit that it is
safe to do even if there is no var associated with the current SN).
This refactor is only possible now that the SN property is set in the
constructor (#1040).