Session: make setCurrentVar fault tolerant
If the current SN doesn't exist in the var data, we need to still
be able to forward to the error.php page. Therefore, we need to make
`setCurrentVar` functional in this case, which we do by not calling
`getCurrentVar` if we know it will fail (i.e. if not `hasCurrentVar`).

As a result, the major change here is that we no longer modify the
previous current var in-place. We replace it, so existing references
to `$var` will no longer be valid.