Fix missing $player definitions
These errors were introduced when the global `$player` was removed
in 867f4fe3.

In cases where we don't have a player (i.e. in the Uni Gen and when
the HoF or Chat Rules is accessed outside of a game), set `$player`
to `null` to avoid a PlayerNotFound exception.

Affected files were found with the following command:

git grep -L "player\s*=" $(git grep -l "(isset|empty).*player")