Remove quasi-global date format constants
Instead of defining some account-dependent date format constants in
`do_voodoo`, just use SmrAccount methods instead. This is much more
straightforward, and we can be sure that they actually exist when
we use them.

The replacements are as follows:

* DATE_DATE_SHORT -> SmrAccount::getDateFormat
* DATE_TIME_SHORT -> SmrAccount::getTimeFormat
* DATE_FULL_SHORT -> SmrAccount::getDateTimeFormat
* DATE_FULL_SHORT_SPLIT -> SmrAccount::getDateTimeFormatSplit

Note that while `getDateTimeFormat` and `getDateTimeFormatSplit` are
new, the others are existing methods that have been renamed:

* {get,set}ShortDateFormat -> {get,set}DateFormat
* {get,set}ShortTimeFormat -> {get,set}TimeFormat

The reason for this is that we don't need the "Short" qualifier since
we are also removing any notion of "Long" date formats, i.e we remove
`DEFAULT_DATE_FULL_LONG` (this was the only "Long" format and was used
sparsely anyway).

As such, we also rename the "DEFAULT" format constants. Here again we
have removed the "FULL" qualifier in favor of the more explicit name
"DATE_TIME" (as we did in the SmrAccount methods):

* DEFAULT_DATE_DATE_SHORT -> DEFAULT_DATE_FORMAT
* DEFAULT_DATE_TIME_SHORT -> DEFAULT_TIME_FORMAT
* DEFAULT_DATE_FULL_SHORT -> DEFAULT_DATE_TIME_FORMAT
* DEFAULT_DATE_FULL_SHORT_SPLIT -> DEFAULT_DATE_TIME_FORMAT_SPLIT